# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Invoice Generation API 
  description: Creates a standardised UBL Invoice from an Order Document and other User-Provided data.
  version: 1.0.0
servers:
    # Added by API Auto Mocking Plugin
    - description: SwaggerHub API Auto Mocking
      url: https://virtserver.swaggerhub.com/unsw-c64/UBL_Invoice_API/1.0.0
    - url: https://api.example.com/api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRegister:
      type: object
      required: 
        - email
        - password
        - businessName
        - abn
      properties:
        email:
          type: string
          format: email
          example: 'testemail@business.com'
        password: 
          type: string
          format: password
          example: 'ILurvc4ts1234!'
        businessName:
          type: string
          example: 'Cuddles4Cats'
        abn:
          type: string
          example: '12345678911'
    RegisterSuccess:
      type: object
      required:
        - adminId
        - token
      properties:
        adminId:
          type: string
          example: 'yeehaw_123'
        token:
          type: string
          description: JWT access token.
          example: 'NJmKsg5oiW...'
    InvoiceId:
      type: integer
      example: 123
    InvoiceIdObj:
      type: object
      properties:
        invoiceId:
          type: integer
          example: 12345
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: 'testemail@business.com'
        password:
            type: string
            format: password
            example: 'ILurvc4ts1234!'
    UserLoginSuccess:
      type: object
      required:
        - adminId
        - token
      properties:
        adminId:
          type: string
          example: 'yeehaw_123'
        token:
          type: string
          description: JWT access token.
          example: 'NJmKsg5oiW...'
    UserUpdate:
      type: object
      required: 
        - email
        - password
        - businessName
      properties:
        email:
          type: string
          format: email
          example: 'newemail@business.com'
        password: 
          type: string
          format: password
          example: 'ILurvc4ts1234!'
        businessName:
          type: string
          example: 'Cuddles4Cats' 
    UserDetails:
      type: object
      required: 
        - email
        - businessName
        - abn
      properties:
        email:
          type: string
          format: email
          example: 'newemail@business.com'
        businessName:
          type: string
          example: 'Cuddles4Cats' 
        abn:
          type: string
          example: '12345678911'
    ErrorResponse:
      type: object
      required: 
        - code
        - error
      properties:
        code:
          type: string
          example: 'ERROR CODE'
        error:
          type: string
          example: 'error message'
    InvoicePeriod:
      type: object
      required:
        - startDate
        - endDate
      properties:
        startDate:
          type: string
          format: date
          description: Start date of the invoicing period
        endDate:
          type: string
          format: date
          description: End date of the invoicing period
    InvoiceData:
      type: object
      required:
        - issueDate
        - dueDate
        - paymentTerms
        - buyer
        - seller
        - lineItems
        - payableAmount
      properties:
        issueDate:
          type: string
          format: date
          description: Date the invoice was issued

        dueDate:
          type: string
          format: date
          description: Date the payment is due

        paymentTerms:
          type: string
          description: Payment terms for the invoice

        InvoicePeriod:
          $ref: '#/components/schemas/InvoicePeriod'

        buyer:
          type: object
          required:
            - name
          properties:
            name:
              type: string
              description: Name of the buyer

        seller:
          type: object
          required:
            - name
          properties:
            name:
              type: string
              description: Name of the seller

        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/orderLines'

        payableAmount:
          type: object
          required:
            - currency
            - amount
          properties:
            currency:
              type: string
              description: Currency code (e.g. AUD, USD)
            amount:
              type: number
              format: float
              description: Total payable amount
    UploadOrderContract:
      type: object
      required: 
        - file
      properties: 
        file:
          type: string
          format: binary
          description: Order contract in XML or JSON.
    names:
      type: object
      required: 
        - name
      properties: 
        name:
          type: string
          example: 'Happy Girl'
    orderLines:
      type: object
      required:
        - lineId
        - itemName
        - quantity
        - unitPrice
        - lineTotal
      properties: 
        lineId:
          type: string
          example: '1'
        itemName:
          type: string
          example: 'Margiela Boots'
        quantity:
          type: number
          example: 2
        unitPrice: 
          type: number
          format: double
          example: 865.00
    ParseOrderContract:
      type: object
      required:
        - orderId
        - buyer
        - seller
        - lines
      properties: 
        orderId:
          type: string
          example: 'OD12345'
        orderDate:
          type: string
          format: date-time
          example: '2026-02-14'
        buyer:
          $ref: '#/components/schemas/names'
        seller:
          $ref: '#/components/schemas/names'
        paymentTerms:
          type: string
          example: 'Payment due within 30 days'
        orderLines:
          type: array
          items:
            $ref: '#/components/schemas/orderLines'
...
    GeneratedInvoice:
      type: object
      required: [invoiceId, invoiceData, invoiceXML]
      properties:
        invoiceId:
          type: string
          example: 'INV13830'
        invoiceStatus:
          type: string
          example: 'draft'
        invoiceData:
          $ref: '#/components/schemas/InvoiceData'
        invoiceXML:
          type: string
          description: generated xml string
          example: "<Invoice xmlns='urn:oasis:names:specification:ubl:schema:xsd:Invoice-2'>...</Invoice>"
    ValidationCheck:
      type: object
      required: 
        - schemaValid
        - businessRulesValid
      properties: 
        schemaValid:
          type: boolean
          example: true
        errors:
          type: array
          items:
            type: string
          example:
            - "Missing required field: issueDate"
            - "Invalid currency code"
paths:
  /v1/admin/auth/register:
    post:
      tags: 
        - 'Account Creation and Management'
      summary: Register a new admin user.
      description: Registers a new admin user using email, password, business name and ABN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses: 
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterSuccess'
        '400':
          description: Invalid email, password or business name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidEmail:
                  summary: Email incorrect format, or email could already exist in database.
                  value:
                    code: 'INVALID_EMAIL'
                    error: 'Email must be a valid email address.'
                weakPassword:
                  summary: Password too short or doesn't contain alphanumeric and symbols.
                  value:
                    code: 'PASSWORD_REJECTED'
                    error: 'Password too weak. Must contain alphanumerics and symbols.'
                invalidABN:
                  summary: ABN less than 11 digits or already exists in database.
                  value:
                    code: 'INVALID_ABN'
                    error: 'ABN must be 11 digits and unregistered in our database.'
  /v1/admin/login:
    post:
      tags: 
        - 'Account Creation and Management'
      summary: Logs in user.
      description: Takes in information about an admin user to determine if they can log in to their account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginSuccess'
        '400':
          description: Failed to login, incorrect email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # /v1/admin/auth/authentication:
  #   get:
  #     tags:
  #       - 'Account Creation and Management'
  #     summary: Verifies user's ABN.
  #     description: Verifies the given ABN using ATO ABN Lookup API
  #     parameters:
  #       - name: abn
  #         in: query
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       "400":
  #         description: Bad request if the ABN is in an invalid format
  #       "404":
  #         description: Not found if ABN is not recognised/registered
  #       "200":
  #         description: ABN is valid
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 status_message:
  #                   type: string
  #                 valid:
  #                   type: boolean
  #                 business_details:
  #                   type: object
  /v1/admin/logout:
    post:
      tags: 
        - 'Account Creation and Management'
      summary: Logs out user and invalidates Token
      security: 
        - BearerAuth: []
      responses: 
        '200':
          description: OK - Logout successful.
          content: 
            application/json:
              schema:
                type: object
        '401':
          description: Token is invalid or empty
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
        '500': 
          description: Internal error (unexpected failure).
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/user/details:
    put:
      tags: 
        - 'Account Creation and Management'
      summary: Update user details 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      security: 
        - BearerAuth: []
      description: Updates user email or business name
      responses: 
        '200': 
          description: Update successful
          content: 
            application/json:
              schema:
                type: object
        '401':
          description: Token is invalid or empty
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
        '500': 
          description: Internal error (unexpected failure).
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
        
    get:
      tags: 
        - 'Account Creation and Management'
      summary: Get the details of an admin user.
      security: 
        - BearerAuth: []
      description: Returns relevant details of loginned user
      responses: 
        '200':
          description: success
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/UserDetails'
        '401':
          description: Token is invalid or empty
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/order/upload:
    post:
      summary: Upload an order document.
      description: Upload a UBL order document in XML or JSON. 
      tags:
        - 'Invoice Generation and Management'
      requestBody: 
        required: true
        content: 
          multipart/form-data: 
            schema: 
              $ref: '#/components/schemas/UploadOrderContract'
      responses: 
        '200':
          description: Order contract upload successful.
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/UploadOrderContract'
        '400':
          description: Invalid file format or Bad Request. 
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/order/parse:
    post:
      summary: Extract order details. 
      description: Extract order details from uploaded Order Contract.
      tags:
        - 'Invoice Generation and Management'
      requestBody: 
        required: true
        content: 
          multipart/form-data: 
            schema:
              $ref: '#/components/schemas/UploadOrderContract'
      responses:
        '200':
          description: successfully parsed Order Document.
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/ParseOrderContract'
        '400':
          description: Invalid format or Bad Request.
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/admin/invoices:
    get:
      tags: 
        - 'Invoice Generation and Management'
      summary: Lists all user's generated UBL invoices.
      security: 
        - BearerAuth: []
      description:
      responses: 
        '200':
          description: Success.
          content:
            application/json: 
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/GeneratedInvoice'
        '401':
          description: Token is invalid or empty
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
        '500': 
          description: Internal error (unexpected failure).
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
  /v1/invoices/{invoiceId}:
    delete:
      tags: 
        - 'Invoice Generation and Management'
      summary: Delete a particular invoice.
      security: 
        - BearerAuth: []
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            $ref: '#/components/schemas/InvoiceId'
      description: Allows a loginned user to generate a standardised UBL Invoice
      responses: 
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
            
        '404':
          description: Token is invalid or empty
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
        '500': 
          description: Internal error (unexpected failure).
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: 
        - 'Invoice Generation and Management'
      summary: Edit a particular invoice.
      security: 
        - BearerAuth: []      
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            $ref: '#/components/schemas/InvoiceId'
      description: Allows a loginned user to edit fields in existing invoice
      responses: 
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceIdObj'
        '404':
          description: Token is invalid or empty
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
        '500': 
          description: Internal error (unexpected failure).
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: 
        - 'Invoice Generation and Management'
      summary: Get a particular invoice
      description: Allows loginned user to retrieve a particular invoice
      security: 
        - BearerAuth: []
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            $ref: '#/components/schemas/InvoiceId'
      responses: 
        '200':
          description: Succesfully retreived invoice
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/GeneratedInvoice'
        '401':
          description: Token is invalid or empty
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: InvoiceId is empty or does not exist
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
        '500': 
          description: Internal error (unexpected failure).
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/invoice/{invoiceId}/xml:
    get:
      summary: Export invoice as XML
      description: Exports the specified invoice as an XML document for the authenticated admin user.
      tags:
        - Invoice Generation and Management
      security:
        - BearerAuth: []
      parameters:
        - name: invoiceId
          in: path
          required: true
          description: ID of the invoice to export
          schema:
            type: string
      responses:
        '200':
          description: Invoice exported successfully as XML
          content:
            application/xml:
              schema:
                type: string
                example: |
                  <?xml version="1.0" encoding="UTF-8"?>
                  <Invoice>
                    <ID>12345</ID>
                    <IssueDate>2026-03-10</IssueDate>
                  </Invoice>
        '400':
          description: Invoice is invalid - does not pass validation against xsd schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/invoices/{invoiceId}/validate:
    post:
      tags:
        - 'Validation'
      summary: Validates generated invoice
      description: Validates the invoice against UBL XSD schema and business rules.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            $ref: '#/components/schemas/InvoiceId'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationCheck'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'